fix : 피드백 반영
1. useEffect 제거, initState로 변경
2. remain.ts -> translate.ts 로 이동
3. useState -> useForm 으로 변경
3. apply 레이아웃 수정
