+++ src/pages/apply/index.tsx
@@ -50,4 +50,5 @@ const Text = styled.p`
   font-weight: 400;
   font-size: 18px;
   line-height: 28px;
+  color: #555555;
 `;
